getting there

+++ apps/mobile-mzima-client/src/app/post/helpers/upload-file-progress.ts
@@ -1,4 +1,4 @@
-import { HttpEvent, HttpEventType, HttpProgressEvent } from '@angular/common/http';
+import { HttpProgressEvent } from '@angular/common/http';
 import { Directory, Filesystem } from '@capacitor/filesystem';
 import { MediaService } from '@mzima-client/sdk';
 import { lastValueFrom, Observable, tap } from 'rxjs';
@@ -16,7 +16,7 @@ export class UploadFileProgressHelper {
       const fetchPhoto = await fetch(data);
       const blob = await fetchPhoto.blob();
       const file = new File([blob], name);
-      const uploadProgress: Observable<HttpEvent<any>> = this.mediaService
+      const uploadProgress: Observable<any> = this.mediaService
         .uploadFileProgress(file, caption)
         .pipe(
           tap((event) => {
@@ -28,27 +28,26 @@ export class UploadFileProgressHelper {
             progressCallback(percentDone);
           }),
         );
-      const event: HttpEvent<any> = await lastValueFrom(uploadProgress);
-      switch (event.type) {
-        case HttpEventType.Response:
-          console.log(`File was completely uploaded!`);
-          for (const content of postData.post_content) {
-            for (const field of content.fields) {
-              if (field.input === 'upload') {
-                field.value.value = event.body.result.id;
-              }
+      const event: any = await lastValueFrom(uploadProgress);
+
+      if (event?.result?.id) {
+        for (const content of postData.post_content) {
+          for (const field of content.fields) {
+            if (field.input === 'upload') {
+              field.value.value = event.result.id;
             }
           }
-          break;
+        }
+
+        Filesystem.deleteFile({
+          directory: Directory.Data,
+          path: path,
+        });
+
+        delete postData.file;
+      } else {
+        throw new Error('Error uploading file!');
       }
-
-      Filesystem.deleteFile({
-        directory: Directory.Data,
-        path: path,
-      });
-
-      delete postData.file;
-      console.log('Postdata - ' + postData);
     } catch (error: any) {
       throw new Error(`Error uploading file: ${error.message}`);
     }
